# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Rimbalzi API
  description: API documentation for Rimbalzi application
  version: 0.1.0
servers:
  - url: https://labinformaticaescuola.it/
    description: Production server
  - url: http://localhost:3000/
    description: Local development server
paths:
...
  /api/session:
    get:
      summary: Get all sessions for a teacher
      parameters:
        - name: uid
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: List of sessions }
        '400': { description: Missing teacher_id }
    post:
      summary: Create a new session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                class_grade: { type: string }
                class_section: { type: string }
                session_phase: { type: string }
                teacher_id: { type: string }
                details: { type: string }
              required: [class_grade, class_section, session_phase, teacher_id]
      responses:
        '200': { description: Ok }
        '400': { description: Invalid input }
    put:
      summary: Update a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teacher_id: { type: string }
                class_id: { type: string }
                session_code: { type: string }
                session_phase: { type: string }
                details: { type: string }
                old_session_phase: { type: string }
              required: [teacher_id, class_id, session_code, session_phase, old_session_phase]
      responses:
        '200': { description: Ok }
        '400': { description: Invalid input }
    delete:
      summary: Delete a session
      parameters:
        - name: session_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Ok }
        '400': { description: Missing or invalid parameters }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []